Data.Var.ST

Description

Documentation

module Data.Var.Class

type STVar = STRef

a value of type STVar s a is a mutable variable in state thread s, containing a value of type a

data STUVar s a

a value of type STUVar s a is a mutable variable in state thread s, containing an unboxed value of type a